+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -1,12 +1,13 @@
 package codegen
 
 object FlutterXPluginInfo {
-    const val Version: String = "5.2.3.242"
+    const val Version: String = "5.3.0.242"
     const val Changelog: String = """
-<h2>5.2.3 - 2025-02-13</h2>
+<h2>5.3.0 - 2025-02-28</h2>
 
-<ul><li>Replaced flutter version update record link</li><li>Fix the wrong command to open an ios project</li><li>Optimize known issues</li></ul>
+<ul><li>New features: Added assets asset path detection mark (dart)</li><li>New features: If a package is marked by the author and stopped updating, it will be marked in the configuration file (
+  yaml)</li><li>New features: Added freezed 3.0 class quick repair tool (dart)</li></ul>
 """
 
-    const val BuildDate: String = "2025-02-13"
+    const val BuildDate: String = "2025-02-28"
 }