+++ src/main/kotlin/shop/itbug/fluttercheckversionx/services/PubspecService.kt
@@ -7,8 +7,20 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.ProjectActivity
+import com.intellij.openapi.vfs.VirtualFileManager
+import com.intellij.openapi.vfs.newvfs.BulkFileListener
+import com.intellij.openapi.vfs.newvfs.events.VFileEvent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
+import shop.itbug.fluttercheckversionx.common.yaml.DartYamlModel
 import shop.itbug.fluttercheckversionx.common.yaml.PubspecYamlFileTools
+import shop.itbug.fluttercheckversionx.util.DartVersionType
 import shop.itbug.fluttercheckversionx.util.MyFileUtil
+import shop.itbug.fluttercheckversionx.util.Version
+import shop.itbug.fluttercheckversionx.util.isVersionGreaterThanThree
+import kotlin.coroutines.CoroutineContext
 
 
 class PubspecStartActivity : ProjectActivity, DumbAware {
@@ -22,12 +34,18 @@ class PubspecStartActivity : ProjectActivity, DumbAware {
  * 仅检测主pubspec.yaml
  */
 @Service(Service.Level.PROJECT)
-class PubspecService(val project: Project) : Disposable {
+class PubspecService(val project: Project) : Disposable, CoroutineScope, BulkFileListener {
 
+    private val job = Job()
     private var dependenciesNames = listOf<String>()
-
+    private var details = listOf<DartYamlModel>()
 
     fun getAllDependencies(): List<String> = dependenciesNames
+
+    init {
+        project.messageBus.connect(parentDisposable = this).subscribe(VirtualFileManager.VFS_CHANGES, this)
+    }
+
 
     /**
      * 读取项目使用了哪些包?
@@ -37,6 +55,7 @@ class PubspecService(val project: Project) : Disposable {
         if (file != null) {
             val tool = PubspecYamlFileTools.create(file)
             val all = tool.allDependencies()
+            details = all
             dependenciesNames = all.map { it.name }
         }
     }
@@ -51,6 +70,19 @@ class PubspecService(val project: Project) : Disposable {
         )
     }
 
+
+    //是否使用了 freezed包
+    fun hasFreezed(): Boolean {
+        return hasDependencies("freezed")
+    }
+
+    //判断 freezed使用的版本号是不是大于 3 或者等于 3
+    fun freezedVersionIsThan3(): Boolean {
+        if (!hasFreezed()) return false
+        val version = details.find { it.name == "freezed" } ?: return false
+        if (version.versionType == DartVersionType.Any) return true
+        return isVersionGreaterThanThree(version.version, Version.parse("3.0.0"))
+    }
 
     /**
      * 项目是否使用 provider 包
@@ -67,6 +99,7 @@ class PubspecService(val project: Project) : Disposable {
         return dependenciesNames.contains(pluginName)
     }
 
+
     companion object {
         fun getInstance(project: Project): PubspecService {
             return project.service<PubspecService>()
@@ -74,8 +107,20 @@ class PubspecService(val project: Project) : Disposable {
     }
 
     override fun dispose() {
-        println("-----dispose--------PubspecService")
+        job.cancel()
         dependenciesNames = emptyList()
+        details = emptyList()
+    }
+
+    override val coroutineContext: CoroutineContext
+        get() = job + Dispatchers.Default
+
+    override fun after(events: List<VFileEvent>) {
+        events.find { it.file?.name == "pubspec.yaml" } ?: return
+        launch(Dispatchers.IO) {
+            startCheck()
+        }
+        super.after(events)
     }
 
 }